Add hamburgr menu to masthead if breadcrumbs disabled. (#2444)
